Fixed Issue#2 (!event join #<4-digits>:
participants variable is now handled as an array. Interpreting an empty
array is left as a task to the function that enumerates the array.

Small fixes:
Added code to announce when a member joins or leaves an event.
This behavior may be disabled by passing the "-silent" argument
when calling !event join or !event leave.